Refactor: extract offlineclient into its own file
