+++ src/test/e2e/services/playground-service.test.ts
@@ -1,7 +1,4 @@
-import {
-    offlineUnleashClient,
-    PlaygroundService,
-} from '../../../lib/services/playground-service';
+import { PlaygroundService } from '../../../lib/services/playground-service';
 import { generateFeatures } from '../../../lib/routes/admin-api/playground.test';
 import { generate as generateContext } from '../../../lib/openapi/spec/sdk-context-schema.test';
 import fc from 'fast-check';
@@ -12,6 +9,7 @@ import FeatureToggleService from '../../../lib/services/feature-toggle-service';
 import { SegmentService } from '../../../lib/services/segment-service';
 import { WeightType } from '../../../lib/types/model';
 import { PlaygroundFeatureSchema } from '../../../lib/openapi/spec/playground-feature-schema';
+import { offlineUnleashClient } from '../../../lib/util/offline-unleash-client';
 
 let stores: IUnleashStores;
 let db: ITestDb;
@@ -155,209 +153,5 @@ describe('the playground service (e2e)', () => {
                 }),
             testParams,
         );
-    });
-});
-
-describe('offline client', () => {
-    it('considers enabled variants with a default strategy to be on', async () => {
-        const name = 'toggle-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    name,
-                    enabled: true,
-                    strategies: [{ name: 'default' }],
-                    variants: [],
-                    type: '',
-                    stale: false,
-                },
-            ],
-            { appName: 'other-app', environment: 'default' },
-            console.log,
-        );
-
-        expect(client.isEnabled(name)).toBeTruthy();
-    });
-
-    it('constrains on appName', async () => {
-        const enabledFeature = 'toggle-name';
-        const disabledFeature = 'other-toggle';
-        const appName = 'app-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    name: enabledFeature,
-                    enabled: true,
-                    strategies: [
-                        {
-                            name: 'default',
-                            constraints: [
-                                {
-                                    contextName: 'appName',
-                                    operator: 'IN',
-                                    values: [appName],
-                                },
-                            ],
-                        },
-                    ],
-                    variants: [],
-                    type: '',
-                    stale: false,
-                },
-                {
-                    name: disabledFeature,
-                    enabled: true,
-                    strategies: [
-                        {
-                            name: 'default',
-                            constraints: [
-                                {
-                                    contextName: 'appName',
-                                    operator: 'IN',
-                                    values: ['otherApp'],
-                                },
-                            ],
-                        },
-                    ],
-                    variants: [],
-                    type: '',
-                    stale: false,
-                },
-            ],
-            { appName, environment: 'default' },
-            console.log,
-        );
-
-        expect(client.isEnabled(enabledFeature)).toBeTruthy();
-        expect(client.isEnabled(disabledFeature)).toBeFalsy();
-    });
-
-    it('considers disabled variants with a default strategy to be off', async () => {
-        const name = 'toggle-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    strategies: [
-                        {
-                            name: 'default',
-                        },
-                    ],
-                    stale: false,
-                    enabled: false,
-                    name,
-                    type: 'experiment',
-                    variants: [],
-                },
-            ],
-            { appName: 'client-test' },
-            console.log,
-        );
-
-        expect(client.isEnabled(name)).toBeFalsy();
-    });
-
-    it('considers disabled variants with a default strategy and variants to be off', async () => {
-        const name = 'toggle-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    strategies: [
-                        {
-                            name: 'default',
-                        },
-                    ],
-                    stale: false,
-                    enabled: false,
-                    name,
-                    type: 'experiment',
-                    variants: [
-                        {
-                            name: 'a',
-                            weight: 500,
-                            weightType: 'variable',
-                            stickiness: 'default',
-                            overrides: [],
-                        },
-                        {
-                            name: 'b',
-                            weight: 500,
-                            weightType: 'variable',
-                            stickiness: 'default',
-                            overrides: [],
-                        },
-                    ],
-                },
-            ],
-            { appName: 'client-test' },
-            console.log,
-        );
-
-        expect(client.isEnabled(name)).toBeFalsy();
-    });
-
-    it("returns variant {name: 'disabled', enabled: false } if the toggle isn't enabled", async () => {
-        const name = 'toggle-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    strategies: [],
-                    stale: false,
-                    enabled: false,
-                    name,
-                    type: 'experiment',
-                    variants: [
-                        {
-                            name: 'a',
-                            weight: 500,
-                            weightType: 'variable',
-                            stickiness: 'default',
-                            overrides: [],
-                        },
-                        {
-                            name: 'b',
-                            weight: 500,
-                            weightType: 'variable',
-                            stickiness: 'default',
-                            overrides: [],
-                        },
-                    ],
-                },
-            ],
-            { appName: 'client-test' },
-
-            console.log,
-        );
-
-        expect(client.isEnabled(name)).toBeFalsy();
-        expect(client.getVariant(name).name).toEqual('disabled');
-        expect(client.getVariant(name).enabled).toBeFalsy();
-    });
-
-    it('returns the disabled variant if there are no variants', async () => {
-        const name = 'toggle-name';
-        const client = await offlineUnleashClient(
-            [
-                {
-                    strategies: [
-                        {
-                            name: 'default',
-                            constraints: [],
-                        },
-                    ],
-                    stale: false,
-                    enabled: true,
-                    name,
-                    type: 'experiment',
-                    variants: [],
-                },
-            ],
-            { appName: 'client-test' },
-
-            console.log,
-        );
-
-        expect(client.getVariant(name, {}).name).toEqual('disabled');
-        expect(client.getVariant(name, {}).enabled).toBeFalsy();
-        expect(client.isEnabled(name, {})).toBeTruthy();
     });
 });
